ResultTools

This is a small module to ease importing Carlo results back into Julia. It contains the function

Carlo.ResultTools.dataframe — Function

ResultTools.dataframe(result_json::AbstractString)

Helper to import result data from a *.results.json file produced after a Carlo calculation. Returns a Tables.jl-compatible dictionary that can be used as is or converted into a DataFrame or other table structure. Observables and their errorbars will be converted to Measurements.jl measurements.

An example of using ResultTools with DataFrames.jl would be the following.

using Plots
using DataFrames
using Carlo.ResultTools

df = DataFrame(ResultTools.dataframe("example.results.json"))

plot(df.T, df.Energy)